とりあえず３つ更新

+++ client/js/flux/index.ts
@@ -1,3 +1,5 @@
+import Todo, { TodoProps } from "../components/todo";
+
 /**
  * Dispatcher
  */
@@ -6,7 +8,7 @@ class Dispatcher extends EventTarget {
     this.dispatchEvent(new CustomEvent("event"));
   }
 
-  subscribe(subscriber) {
+  subscribe(subscriber: any) {
     this.addEventListener("event", subscriber);
   }
 }
@@ -14,6 +16,12 @@ class Dispatcher extends EventTarget {
 /**
  * Action Creator and Action Types
  */
+
+type Action = {
+  type: string,
+  payload: TodoProps | undefined
+};
+
 const FETCH_TODO_ACTION_TYPE = "Fetch todo list from server";
 export const createFetchTodoListAction = () => ({
   type: FETCH_TODO_ACTION_TYPE,
@@ -21,19 +29,19 @@ export const createFetchTodoListAction = () => ({
 });
 
 const ADD_TODO_ACTION_TYPE = "A todo addition to store";
-export const createAddTodoAction = todo => ({
+export const createAddTodoAction = (todo: TodoProps) => ({
   type: ADD_TODO_ACTION_TYPE,
   payload: todo
 });
 
 const UPDATE_TODO_ACTION_TYPE = "Update todo state";
-export const updateTodoAction = todo => ({
+export const updateTodoAction = (todo: TodoProps) => ({
   type: UPDATE_TODO_ACTION_TYPE,
   payload: todo
 });
 
 const REMOVE_TODO_ACTION_TYPE = "Remove todo";
-export const removeTodoAction = todo => ({
+export const removeTodoAction = (todo: TodoProps) => ({
   type: REMOVE_TODO_ACTION_TYPE,
   payload: todo
 });
@@ -58,8 +66,8 @@ const headers = {
   "Content-Type": "application/json; charset=utf-8"
 };
 
-const reducer = async (prevState, { type, payload }) => {
-  switch (type) {
+const reducer = async (prevState: any, action: Action) => {
+  switch (action.type) {
     case FETCH_TODO_ACTION_TYPE: {
       try {
         const resp = await fetch(api).then(d => d.json());
@@ -69,7 +77,11 @@ const reducer = async (prevState, { type, payload }) => {
       }
     }
     case UPDATE_TODO_ACTION_TYPE: {
-      const { id, ...body } = payload;
+      if(action.payload==null) {
+        console.log("payloadがnullです");
+        return;
+      }
+      const { id, ...body } = action.payload;
       try {
         const resp = await fetch(`${api}/${id}`, {
           method: "PATCH",
@@ -79,7 +91,7 @@ const reducer = async (prevState, { type, payload }) => {
           },
           body: JSON.stringify(body)
         }).then(d => d.json());
-        const idx = prevState.todoList.findIndex(todo => todo.id === resp.id);
+        const idx = prevState.todoList.findIndex((todo: TodoProps) => todo.id === resp.id);
         if (idx === -1) return prevState;
         const nextTodoList = prevState.todoList.concat();
         nextTodoList[idx] = resp;
@@ -89,13 +101,17 @@ const reducer = async (prevState, { type, payload }) => {
       }
     }
     case REMOVE_TODO_ACTION_TYPE: {
-      const { id } = payload;
+      if(action.payload==null) {
+        console.log("payloadがnullです");
+        return;
+      }
+      const id = action.payload.id;
       try {
         await fetch(`${api}/${id}`, {
           method: "DELETE",
           mode: "cors"
         });
-        const idx = prevState.todoList.findIndex(todo => todo.id == id);
+        const idx = prevState.todoList.findIndex((todo: TodoProps) => todo.id == id);
         if (idx === -1) return prevState;
         const nextTodoList = prevState.todoList.concat();
         nextTodoList.splice(idx, 1);
@@ -105,7 +121,11 @@ const reducer = async (prevState, { type, payload }) => {
       }
     }
     case ADD_TODO_ACTION_TYPE: {
-      const body = JSON.stringify(payload);
+      if(action.payload==null) {
+        console.log("payloadがnullです");
+        return;
+      }
+      const body = JSON.stringify(action.payload);
       const config = { method: "POST", body, headers };
       try {
         const resp = await fetch(api, config).then(d => d.json());
@@ -118,7 +138,7 @@ const reducer = async (prevState, { type, payload }) => {
       return { ...prevState, error: null };
     }
     default: {
-      throw new Error("unexpected action type: %o", { type, payload });
+      throw new Error("unexpected action type");
     }
   }
 };
@@ -127,16 +147,16 @@ export function createStore(initialState = defaultState) {
   const dispatcher = new Dispatcher();
   let state = initialState;
 
-  const dispatch = async ({ type, payload }) => {
-    console.group(type);
+  const dispatch = async (action: Action) => {
+    console.group(action.type);
     console.log("prev", state);
-    state = await reducer(state, { type, payload });
+    state = await reducer(state, action);
     console.log("next", state);
     console.groupEnd();
     dispatcher.dispatch();
   };
 
-  const subscribe = subscriber => {
+  const subscribe = (subscriber: any) => {
     dispatcher.subscribe(() => subscriber(state));
   };
